fix(billing): cast to numeric so the near-limit threshold param binds (operacional 500)
getOperationalHealth interpolated NEAR_LIMIT_THRESHOLD (0.8) into a raw
tx.execute as `uc.conversas_limite * ${0.8}`. Postgres infers the bound param's
type from `integer * $1` as integer, so postgres.js sending "0.8" failed with
`22P02 invalid input syntax for type integer: "0.8"` — 500ing the super-admin
Operacional dashboard for every load, data-independent (bind fails before
execution). Same mock-blind class as the analytics Date bug.

Fix: cast the left operand `uc.conversas_limite::numeric` so the param is
inferred as numeric and 0.8 binds. Verified: getOperationalHealth returns OK and
the /operacional dashboard renders (was 500).

Found during usability-testing session (roteiro J-12 / F-26).

+++ packages/billing/src/use-cases/get-operational-health.ts
@@ -129,7 +129,7 @@ export async function getOperationalHealth(usdToBrlRate: number): Promise<Operat
       ) owner ON true
       WHERE uc.periodo = TO_CHAR(CURRENT_DATE, 'YYYY-MM')
         AND uc.conversas_limite > 0
-        AND uc.conversas_usadas >= uc.conversas_limite * ${NEAR_LIMIT_THRESHOLD}
+        AND uc.conversas_usadas >= uc.conversas_limite::numeric * ${NEAR_LIMIT_THRESHOLD}
       ORDER BY usage_pct DESC
     `)) as Array<{
       tenant_id: string;
